// Frontend/src/models/LnkMetaData.ts
export interface LnkMetaData {
  ShortcutPath: string;
  TargetPath: string;
  WorkingDirectory: string;
  Arguments: string;
  Description: string;
  IconPath: string;
  IconIndex: number;
  CreationTime: string;
  LastModifiedTime: string;
  LastAccessedTime: string;
  IsTargetValid: boolean;
  TargetMimeType: string;
  TargetExtension: string;
  DisplayName: string;
}